+++ cmd/block-writer/main.go
@@ -16,7 +16,6 @@ package main
 import (
 	"context"
 	"flag"
-	"log"
 
 	// use mysql
 	_ "github.com/go-sql-driver/mysql"
@@ -44,14 +43,9 @@ func main() {
 		RunTime:     fixture.Context.RunTime,
 		RunRound:    1,
 	}
-
-	provisioner, err := cluster.NewK8sProvisioner()
-	if err != nil {
-		log.Fatal(err)
-	}
 	suit := util.Suit{
 		Config:        &cfg,
-		Provisioner:   provisioner,
+		Provisioner:   cluster.NewK8sProvisioner(),
 		ClientCreator: blockwriter.CaseCreator{TableNum: *tables, Concurrency: *concurrency},
 		NemesisGens:   util.ParseNemesisGenerators(fixture.Context.Nemesis),
 		VerifySuit:    verify.Suit{},
