+++ cmd/on-dup/main.go
@@ -16,7 +16,6 @@ package main
 import (
 	"context"
 	"flag"
-	"log"
 
 	// use mysql
 	_ "github.com/go-sql-driver/mysql"
@@ -45,15 +44,9 @@ func main() {
 		RunTime:     fixture.Context.RunTime,
 		RunRound:    1,
 	}
-
-	provisioner, err := cluster.NewK8sProvisioner()
-	if err != nil {
-		log.Fatal(err)
-	}
-
 	suit := util.Suit{
 		Config:      &cfg,
-		Provisioner: provisioner,
+		Provisioner: cluster.NewK8sProvisioner(),
 		ClientCreator: ondup.CaseCreator{Cfg: &ondup.Config{
 			DBName:     *dbName,
 			NumRows:    *numRows,
