--- docker/.gitignore
@@ -1,3 +1,0 @@
-secret/*
-!.gitkeep
-
