--- docker/docker-compose.dev.yml
@@ -1,6 +1,0 @@
-version: '3.7'
-services:
-  control:
-    volumes:
-      # Mounts $CHAOS_ROOT on host to /go/src/github.com/pingcap/tipocket control container
-      - ${CHAOS_ROOT}:/go/src/github.com/pingcap/tipocket
