--- docker/docker-compose.yml
@@ -1,43 +1,0 @@
-version: '3.7'
-x-node:
-  &default-node
-  build: ./node
-  env_file: ./secret/node.env
-  privileged: true
-  networks:
-    - chaos
-
-services:
-  control:
-    container_name: chaos-control
-    hostname: control
-    build: ./control
-    env_file: ./secret/control.env
-    privileged: true
-    ports:
-      - "8080"
-    networks:
-      - chaos
-  n1:
-    << : *default-node
-    container_name: chaos-n1
-    hostname: n1
-  n2:
-    << : *default-node
-    container_name: chaos-n2
-    hostname: n2
-  n3:
-    << : *default-node
-    container_name: chaos-n3
-    hostname: n3
-  n4:
-    << : *default-node
-    container_name: chaos-n4
-    hostname: n4
-  n5:
-    << : *default-node
-    container_name: chaos-n5
-    hostname: n5
-
-networks:
-  chaos:
