--- jobs/bank_1.yaml
@@ -1,47 +1,0 @@
----
-apiVersion: batch/v1beta1
-kind: CronJob
-metadata:
-  name: bank-1
-  namespace: tipocket-job
-spec:
-  schedule: "0 */4 * * *"
-  concurrencyPolicy: Forbid
-  jobTemplate:
-    spec:
-      template:
-        spec:
-          containers:
-          - name: tipocket
-            imagePullPolicy: Always
-            image: pingcap/tipocket:latest
-            command:
-              - "sh"
-              - "-c"
-              - |
-                /bin/chaos-tidb \
-                -case=bank \
-                -namespace=tipocket-test \
-                -image-version=v3.0.9 \
-                -client=5 \
-                -hub=hub.pingcap.net \
-                -nemesis=random_kill,partition_one,short_kill_tikv_1node,short_kill_pd_leader,kill_tikv_1node_5min \
-                -round=3 \
-                -request-count=10000
-
-                if [ $? != "0" ]; then
-                  echo "ERROR OCCURED, PTAL..."
-                  while true; do sleep 60; done
-                fi
-            env:
-            - name: KUBECONFIG
-              value: "/etc/config/kube_config"
-            volumeMounts:
-            - name: kube-config
-              mountPath: /etc/config
-          volumes:
-          - name: kube-config
-            configMap:
-              name: kube-config
-          restartPolicy: Never
-      backoffLimit: 0
