+++ pkg/util/pdutil/client.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/pingcap/errors"
 	"github.com/pingcap/kvproto/pkg/metapb"
+
 	httputil "github.com/pingcap/tipocket/pkg/util/http"
 )
