+++ tests/pessimistic/deadlock/deadlock.go
@@ -10,6 +10,7 @@ import (
 
 	"github.com/ngaut/log"
 	"github.com/pingcap/errors"
+
 	"github.com/pingcap/tipocket/util"
 )
