+++ tests/pessimistic/deadlock/leader_change.go
@@ -7,6 +7,7 @@ import (
 	"time"
 
 	"github.com/ngaut/log"
+
 	"github.com/pingcap/tipocket/pkg/cluster/types"
 	"github.com/pingcap/tipocket/pkg/core"
 )
